readline: Compensate for odd scrolling behavior.
The simics console won't scroll when at the end of the screen until the
first character of the new line is typed.  This confuses the readline
console, which tracks where the cursor should be.  To compensate, detect
when potentially in this situation, and output a space followed by a
backspace.

Signed-off-by: Scott Wood <scottwood@freescale.com>