printlog() improvements
Allow LOGLEVEL_ALWAYS to be used without a silly warning.

Optimize for the not-printed case (it will be correct for the more
verbose loglevels, and less verbose loglevel prints should be
infrequent enough that it doesn't matter).

Don't bother with the max build loglevel comparison if the loglevel is
not a constant (unlikely, but legitimate).

Add LOGTYPE_MALLOC.

Signed-off-by: Scott Wood <scottwood@freescale.com>