libos-client.h: Add virt_to_phys() and PHYSMAPSIZE.
virt_to_phys() gives the client the ability to translate virtual
addresses into physical, without libos assuming that the PHYSBASE mapping
is still present and still points to zero.  Clients should #define
HAVE_VIRT_TO_PHYS to make use of this.

PHYSMAPSIZE gives the client the ability to specify the size
of the PHYSBASE mapping.  It defaults to 256MiB if not specified.

The example libos-client.h is updated.

Signed-off-by: Scott Wood <scottwood@freescale.com>