Make "crashing" bypass locks on console output path only.
"crashing" is now a CPU-local variable that skips console-path locks on
the crashing core, but do not try to recurse more than once through
printf().  This makes crashes from within the console subsystem more
likely to be printed properly.

Remove puts_len() (it's not used), and reorganize puts() and putchar(),
to reduce duplication of code.

Signed-off-by: Scott Wood <scottwood@freescale.com>