Optimize memset() for common cases.
When the character to set is zero, and the buffer is word-aligned,
use dcbz and word-stores when possible.

With this, gcc2.sh performs better with the TLB cache than without
(approx 5 million fewer supervisor instructions and 11 million fewer
hypervisor instructions).

gcc1.sh performs substantially better with the TLB cache (10 million
supervisor versus 14.7 million without the cache, and 183 million
hypervisor versus 540 million without the cache).

find.sh performs a little worse with the TLB cache (6.77 million
hypervisor versus 4.68 million without the cache), but nowhere near as
bad as before, when the TLB cache multiplied the hypervisor instructions
by a factor of 5.

Signed-off-by: Scott Wood <scottwood@freescale.com>