Add extra compile-time checking.
If C=1 is passed to make, then sparse (via the cgcc wrapper) is used
to issue additional warnings (these warnings are not turned into
errors, unlike GCC warnings).  This is like Linux.

If C=2 is passed, then in addition to running sparse, extra warnings
are enabled.  These are warnings that are prone to false positives,
so they are not normally enabled and -Werror is turned off in this mode.

This is unlike Linux, which uses C=2 to mean that all files are checked
regardless of whether they need rebuilding (use make clean for that instead).

This patch fixes many issues shown by these warnings, though there are
some remaining issues with sparse in that it is missing certain symbols
that GCC defines.

There are some changes to the libos client interface (start was renamed
to avoid namespace pollution, and the makefiles interact a little
differently), so pull this subproject commit into the client's git
atomically with updating the client.

Signed-off-by: Scott Wood <scottwood@freescale.com>