Use GCC intrinsics for zero-counting.
This removes a 64-bit FIXME, and allows invocations with constant
arguments to produce constant results.

Note that count_msb_zeroes no longer produces defined results when passed
zero.  This brings it in line with the other functions and the GCC
semantics.  I only saw one call site (in fh_send_nmi) which did not check
for zero; it will be fixed.

Signed-off-by: Scott Wood <scottwood@freescale.com>