queue: Fix blocking behavior.
The blocking versions of write and writechar were not using the API
properly, and were notifying the consumer immediately before blocking.

If the consumer drained the entire queue during that callback, then the
caller would never wake up.

Also, the consumer would fail to be notified until the queue were full.

Signed-off-by: Scott Wood <scottwood@freescale.com>