Libos changes for HV Error queue implementation.
This patch provides an event queue mechanism for
dumping platform machine check information
information. The machine check data (recoverable
machine check condition) is stored in a
queue and an event is set on for the current gcpu
while machine check is being serviced.  The data is
dumped to the cosole when the event is serviced at a later point.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>