libos: remove decrementer handler in latency test
The decrementer ISR was being called because there was a pending decrementer
interrupt when interrupts were enabled.  If we clear the DIS bit first, then
the decrementer will never generate an interrupt, so we can get rid of the
interrupt handler

Also fix the vector value used for the MPIC programming.

Signed-off-by: Timur Tabi <timur@freescale.com>