Support booting from non-zero addresses.
The boot code now properly treats the device tree pointer as an effective
address, translating to physical it in early boot code.  While doing so,
it remembers the offset between physical and IMA effective.  Later, this
offset is used to turn PHYSBASE into a physical address.  This physical
address is saved for later use by virt_to_phys().

The IMA size was being used improperly -- it was assumed to be the size of
the actual TLB entry, but ePAPR allows the IMA to consist of multiple TLB
entries, and an IMA TLB entry can cover more than just the IMA.  The IMA
size is no longer used at all.

Signed-off-by: Scott Wood <scottwood@freescale.com>