latency: disable all core timer interrupts
This is a more robust solution than clearing DIS.

Also, I've observed a write to TSR to clear DIS not preventing the
interrupt from being delivered if the very next thing is enabling EE.
An isync makes that go away, even though it's not listed in the
synchronization requirements section of the e500 manual.

Clearing TCR doesn't appear to need an isync.

Signed-off-by: Scott Wood <scottwood@freescale.com>