Added tsize_to_pages_roundup function.
The minimum translation size supported is 4K and the tsize_to_pages
function returnes 0 pages for an unsupported low value of tsize.
However, in case of page table walk, the page table size is allowed
to be 1K or 2K. tsize_to_pages_roundup function was added to support
this scenarion returning one page for 1K and 2K sizes.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>