T4: Change config option to signify number of CPUs not cores
Replace LIBOS_MAX_CORES that counted the number of cores
(a core is defined as an ensemble of several threads,
common hw blocks etc.) by a LIBOS_MAX_CPUS config.
This new config counts the maximum number of supported
CPUs and *includes* hardware threads.

E.g. For a T4240 SoC LIBOS_MAX_CORES was 12 and
the new LIBOS_MAX_CPUS is 24.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>