T4: Increase number of supported internal irqs
Increase the maximum number of supported internal
irqs to 512 even if the new mpic supports only
256 sources. That's because the irq numbers in
the T4 hw devicetree were tweaked so that the
irqnums > 128 fit right on their register map range.
Or, in other words, these irqnums were made larger
than 256 because the irqs no longer map linearly
with the registers.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>