Let caller specify IPROT on tlb1 entries
Modify set_tlb1_entry() so that instead of
silently setting IPROT on the newly added tlb1
entry, let the caller choose whether it wants
IPROT on or off.
In order to do this, modify the set_tlb1_entry()
prototype by adding a mas1flags parameter, along
the other masXflags parameters. This has the
added benefit that the prototype is simplified by
dropping two parameters: "ts" and "indirect".
Also in this patch:
 * refactored a bit set_tlb1_entry() by dropping
   a couple of local vars.
 * updated call sites

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>