workaround potential issue with default cpu features on MMUv2
In case the cpu features table wasn't updated with the
PVR we're running on we'll end up with the default feature mask.
This might mean a mismatch between the actual MMUv2 presence and
its flag not being set in the cpu features mask. This will cause
weird crashes further in the code, e.g. because TLBxCFG is parsed
using MMUv1 format instead of MMUv2.

This code runs very early, not even the IVORs are set up, so the
obvious solution of using an assert() doesn't work.
Instead hot fix the cpu features so execution can continue
and add a big warning that will display first as hv starts.

To replicate, comment all but the default entry in cpu features table.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>